Move read and write functionality to corresponding data structures
