+++ src/main/java/com/kichik/pecoff4j/ImportDirectory.java
@@ -9,9 +9,11 @@
  *******************************************************************************/
 package com.kichik.pecoff4j;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.kichik.pecoff4j.io.DataReader;
 import com.kichik.pecoff4j.util.DataObject;
 
 public class ImportDirectory extends DataObject {
@@ -19,6 +21,29 @@ public class ImportDirectory extends DataObject {
 	private List<String> names = new ArrayList();
 	private List<ImportDirectoryTable> nameTables = new ArrayList();
 	private List<ImportDirectoryTable> addressTables = new ArrayList();
+
+	public static ImportDirectory read(byte[] b, int baseAddress)
+			throws IOException {
+		DataReader dr = new DataReader(b);
+		ImportDirectory id = new ImportDirectory();
+		ImportDirectoryEntry ide = null;
+		while ((ide = ImportDirectoryEntry.read(dr)) != null) {
+			id.add(ide);
+		}
+
+		/*
+		 * FIXME - name table refer to data outside image directory for (int i =
+		 * 0; i < id.size(); i++) { ImportDirectoryEntry e = id.getEntry(i);
+		 * dr.jumpTo(e.getNameRVA() - baseAddress); String name = dr.readUtf();
+		 * dr.jumpTo(e.getImportLookupTableRVA() - baseAddress);
+		 * ImportDirectoryTable nt = readImportDirectoryTable(dr, baseAddress);
+		 * dr.jumpTo(e.getImportAddressTableRVA() - baseAddress);
+		 * ImportDirectoryTable at = null; // readImportDirectoryTable(dr, //
+		 * baseAddress); id.add(name, nt, at); }
+		 */
+
+		return id;
+	}
 
 	public void add(ImportDirectoryEntry entry) {
 		entries.add(entry);
