+++ src/main/java/com/kichik/pecoff4j/PESignature.java
@@ -9,12 +9,28 @@
  *******************************************************************************/
 package com.kichik.pecoff4j;
 
+import com.kichik.pecoff4j.io.IDataReader;
+import com.kichik.pecoff4j.io.IDataWriter;
+
+import java.io.IOException;
 import java.util.Arrays;
 
 public class PESignature {
 	private static byte[] expected1 = new byte[] { 0x50, 0x45, 0x00, 0x00 };
 	private static byte[] expected2 = new byte[] { 0x50, 0x69, 0x00, 0x00 };
 	private byte[] signature;
+
+	public static PESignature read(IDataReader dr) throws IOException {
+		PESignature ps = new PESignature();
+		byte[] signature = new byte[4];
+		dr.read(signature);
+		ps.setSignature(signature);
+		return ps;
+	}
+
+	public void write(IDataWriter dw) throws IOException {
+		dw.writeBytes(getSignature());
+	}
 
 	public byte[] getSignature() {
 		return signature;
