+++ src/main/java/com/kichik/pecoff4j/ResourceDirectory.java
@@ -9,14 +9,28 @@
  *******************************************************************************/
 package com.kichik.pecoff4j;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.kichik.pecoff4j.io.IDataReader;
 import com.kichik.pecoff4j.util.DataObject;
 
 public class ResourceDirectory extends DataObject {
 	private ResourceDirectoryTable table;
 	private List<ResourceEntry> entries = new ArrayList();
+
+	public static ResourceDirectory read(IDataReader dr, int baseAddress) throws IOException {
+		ResourceDirectory d = new ResourceDirectory();
+		d.setTable(ResourceDirectoryTable.read(dr));
+		int ne = d.getTable().getNumNameEntries()
+				+ d.getTable().getNumIdEntries();
+		for (int i = 0; i < ne; i++) {
+			d.add(ResourceEntry.read(dr, baseAddress));
+		}
+
+		return d;
+	}
 
 	public ResourceDirectoryTable getTable() {
 		return table;
