+++ src/main/java/com/kichik/pecoff4j/ResourceDirectoryTable.java
@@ -9,6 +9,10 @@
  *******************************************************************************/
 package com.kichik.pecoff4j;
 
+import com.kichik.pecoff4j.io.IDataReader;
+
+import java.io.IOException;
+
 public class ResourceDirectoryTable {
 	private int characteristics;
 	private int timeDateStamp;
@@ -16,6 +20,18 @@ public class ResourceDirectoryTable {
 	private int minVersion;
 	private int numNameEntries;
 	private int numIdEntries;
+
+	public static ResourceDirectoryTable read(IDataReader dr) throws IOException {
+		ResourceDirectoryTable t = new ResourceDirectoryTable();
+		t.setCharacteristics(dr.readDoubleWord());
+		t.setTimeDateStamp(dr.readDoubleWord());
+		t.setMajorVersion(dr.readWord());
+		t.setMinVersion(dr.readWord());
+		t.setNumNameEntries(dr.readWord());
+		t.setNumIdEntries(dr.readWord());
+
+		return t;
+	}
 
 	public int getCharacteristics() {
 		return characteristics;
