--- src/main/java/com/kichik/pecoff4j/io/SectionAssembler.java
@@ -1,14 +1,0 @@
-/*******************************************************************************
- * This program and the accompanying materials
- * are made available under the terms of the Common Public License v1.0
- * which accompanies this distribution, and is available at 
- * http://www.eclipse.org/legal/cpl-v10.html
- * 
- * Contributors:
- *     Peter Smith
- *******************************************************************************/
-package com.kichik.pecoff4j.io;
-
-public class SectionAssembler {
-
-}
